+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Unreleased
 
+NOTES:
+
+* Update vRO payload.
+
+FEATURES:
+
+* **Execution Request** Client methods for `guidance_settings` API endpoints
 
 ## 0.2.10 (March 1, 2023)